Cleanup, refactoring and documentation.
