+++ promotion/src/main/java/org/cryptimeleon/incentive/promotion/promotions/NutellaPromotion.java
@@ -1,17 +1,13 @@
 package org.cryptimeleon.incentive.promotion.promotions;
 
 import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.Value;
 import lombok.experimental.NonFinal;
 import org.cryptimeleon.incentive.crypto.IncentiveSystem;
 import org.cryptimeleon.incentive.crypto.model.PromotionParameters;
 import org.cryptimeleon.incentive.promotion.model.Basket;
 import org.cryptimeleon.incentive.promotion.model.BasketItem;
-import org.cryptimeleon.incentive.promotion.reward.NutellaReward;
 import org.cryptimeleon.incentive.promotion.reward.Reward;
-import org.cryptimeleon.math.serialization.ListRepresentation;
-import org.cryptimeleon.math.serialization.ObjectRepresentation;
 import org.cryptimeleon.math.serialization.Representation;
 import org.cryptimeleon.math.serialization.annotations.ReprUtil;
 import org.cryptimeleon.math.serialization.annotations.Represented;
@@ -19,8 +15,6 @@ import org.cryptimeleon.math.structures.cartesian.Vector;
 
 import java.math.BigInteger;
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 /**
  * This promotion is a classic promotion from the paper:
@@ -55,12 +49,6 @@ public class NutellaPromotion implements Promotion {
                         .reduce(Integer::sum)
                         .orElseThrow()
         ));
-    }
-
-    public List<Reward> computeRewardsForPoints(Vector<BigInteger> tokenPoints, Vector<BigInteger> basketPoints) {
-        return this.rewards.stream()
-                .filter(reward -> reward.computeSatisfyingNewPointsVector(tokenPoints, basketPoints).isPresent())
-                .collect(Collectors.toList());
     }
 
     public Representation getRepresentation() {
