+++ promotion/src/test/java/org/cryptimeleon/incentive/promotion/promotions/NutellaPromotionTest.java
@@ -1,7 +1,5 @@
 package org.cryptimeleon.incentive.promotion.promotions;
 
-import org.cryptimeleon.incentive.crypto.Setup;
-import org.cryptimeleon.incentive.crypto.model.IncentivePublicParameters;
 import org.cryptimeleon.incentive.crypto.model.PromotionParameters;
 import org.cryptimeleon.incentive.promotion.reward.NutellaReward;
 import org.cryptimeleon.incentive.promotion.reward.Reward;
@@ -17,7 +15,6 @@ class NutellaPromotionTest {
 
     @Test
     void representationTest() {
-        IncentivePublicParameters publicParameters = Setup.trustedSetup(128, Setup.BilinearGroupChoice.Debug);
         PromotionParameters promotionParameters = NutellaPromotion.generatePromotionParameters();
         List<Reward> rewardList = List.of(new NutellaReward(4, UUID.randomUUID(), new RewardSideEffect("Free Nutella")));
         NutellaPromotion nutellaPromotion = new NutellaPromotion(promotionParameters, rewardList);
