+++ promotion/src/test/java/org/cryptimeleon/incentive/promotion/reward/RewardSideEffectTest.java
@@ -12,5 +12,4 @@ class RewardSideEffectTest {
         RewardSideEffect deserializedRewardSideEffect = new RewardSideEffect(rewardSideEffect.getRepresentation());
         assertEquals(rewardSideEffect, deserializedRewardSideEffect);
     }
-
 }
